Down to 48 regs for torsion grad, a few more tweaks

+++ src/forcefields/mmff.h
@@ -72,9 +72,9 @@ struct TorsionContribTerms {
   std::vector<int>    idx2;
   std::vector<int>    idx3;
   std::vector<int>    idx4;
-  std::vector<double> V1;
-  std::vector<double> V2;
-  std::vector<double> V3;
+  std::vector<float> V1;
+  std::vector<float> V2;
+  std::vector<float> V3;
 };
 
 struct VdwTerms {
@@ -180,9 +180,9 @@ struct TorsionContribTermsDevice {
   nvMolKit::AsyncDeviceVector<int>    idx2;
   nvMolKit::AsyncDeviceVector<int>    idx3;
   nvMolKit::AsyncDeviceVector<int>    idx4;
-  nvMolKit::AsyncDeviceVector<double> V1;
-  nvMolKit::AsyncDeviceVector<double> V2;
-  nvMolKit::AsyncDeviceVector<double> V3;
+  nvMolKit::AsyncDeviceVector<float> V1;
+  nvMolKit::AsyncDeviceVector<float> V2;
+  nvMolKit::AsyncDeviceVector<float> V3;
 };
 
 struct VdwTermsDevice {
